Types refactoring
